完成pretty-code命令执行 - 核心引擎系统初始化模块优化
优化了核心引擎系统初始化和配置管理模块：
- pretty/02_core_engine_part003.c: 系统初始化、配置管理、数据处理

+++ TaleWorlds.Native/pretty/02_core_engine_part003.c
@@ -802,7 +802,7 @@ void core_engine_system_initializer_type13(void)
     // 设置系统注册信息
     parent_node[6] = SYSTEM_REGISTRY_ID_11;
     parent_node[7] = SYSTEM_REGISTRY_ID_12;
-    parent_node[8] = &global_var_7584_ptr;
+    parent_node[8] = (uint64_t)&global_var_7584_ptr;
     parent_node[9] = 1;
     parent_node[10] = callback_function;
     
@@ -1264,7 +1264,7 @@ void core_engine_system_initializer_type20(void)
     // 设置系统注册信息
     parent_node[6] = SYSTEM_REGISTRY_ID_11;
     parent_node[7] = SYSTEM_REGISTRY_ID_12;
-    parent_node[8] = &global_var_7584_ptr;
+    parent_node[8] = (uint64_t)&global_var_7584_ptr;
     parent_node[9] = 1;
     parent_node[10] = callback_function;
